An initial test implementation of training.

+++ DeepNetwork.Infrastructure/NetworkTrainer/NetworkTrainer.h
@@ -8,12 +8,12 @@
 class NetworkTrainer
 {
 private:
-	std::vector<float> dError_dLayerAbove;
-	std::vector<float> dError_dOutputCurrent;
-	std::vector<std::vector<float>> adjustmentsStorage;
+	std::vector<double> dError_dLayerAbove;
+	std::vector<double> dError_dOutputCurrent;
+	std::vector<std::vector<double>> adjustmentsStorage;
 	std::unique_ptr<Logger> logger;
 
-	float CalculateErrorDerivativeForFinalLayer(matrix finalLayer, matrix expectedLayer);
+	double CalculateErrorDerivativeForFinalLayer(matrix finalLayer, matrix expectedLayer);
 	void UpdateWeights(network network);
 	void GetAdjustmentsForLayer(network network, int a);
 	void GetAdjustments(network network);
@@ -21,6 +21,6 @@ private:
 	void GetErrorDerivativeForOutputLayer(matrix weightMatrix, matrix outputLayer);
 public:
 	NetworkTrainer(network network);
-	float TrainNetwork(network network, matrix expectedLayer);
+	double TrainNetwork(network network, matrix expectedLayer);
 };
