+++ DeepNetwork.Infrastructure/Train Network/train_network.h
@@ -5,7 +5,7 @@
 extern "C" {
 #define EXPORT_SYMBOL __declspec(dllexport)
 
-	EXPORT_SYMBOL float train_network(network network, matrix expectedLayer);
+	EXPORT_SYMBOL double train_network(network network, matrix expectedLayer);
 
 #undef EXPORT_SYMBOL
 }
