+++ DeepNetwork/matrix.py
@@ -1,7 +1,7 @@
 from ctypes import *
 
 class matrix(Structure):
-    _fields_ = [("values", POINTER(c_float)),
+    _fields_ = [("values", POINTER(c_double)),
                 ("rows", c_int),
                 ("cols", c_int)]
